Final Section 1
Completed the first 1 of the project, where it is to find the information from the website, and transforming the function responsible of this, in such a way that now can recieve a page source to calculate those findings.
HINT: The page source will be provided  after loading all the page with selenium.